Change scServer to scOutput and graph changes
Now we return to the old way of having a node that can choose the server and the channel to output the input bus. Now multiple instances are allowed, and there is more flexibility in output routing in this manner. While doing so, changes in how the graph is computed have been made, so a big rewrite to serverManager::recomputeGraph method.