Update metadata for in, bufs and outs and multiout
Now inputs, buffer parameters and outputs are defined via ControlSpecs not global metadata tags. This allows more flexibility with the names, and also we added the bonus of multiple outputs per synthdef.
The main change is now the connection within nodes are nodePorts, that store the reference to the bus and also which port is, indexed by [0...n-1]