Add parameter modulation with an audio bus
With an special declaration in the synthdef, and some special parameters (name_sel and name_ar) a parameter can be controlled with a control rate signal, or with an audio rate signal comming from an audio bus.
This changed also afected graph calculation, as now every parameter that can be audio modulated, in graph prespective, is a possible input.
See scSynthdefs.sc for example on how to use it, and be aware of the pseudo-ugen that needs to be declared.
Check last changes in ofxSupercollider for mapa and mapan methos