[Variables] Define vars that reload the synthdef
You can define n variables that will be published in the node, and when changed make the synth being recreated.
This are meant for parameters than need to be specified when building the synthdef, in the panner example, the number of speakers has to be a fixed value when creating the PanX Ugen.
The synthdef name will be in the form of {name}n_x_y_z_...
Where name is the name of the synthdef.
n is the number of voices
x is the first paramter
y is the second parameter
z is the third
etc...